analog voltage addition works, or appears to
Added AnalogInputVoltage functionality. It occurs to me that when a task
returns its channels, unless i happen to have created that class, I
cannot utilize the information. So, I need to add the ability to "fake"
create my class from the base down instead of the normal way around for
the case that the class exists and I do not need to call the C-API.

The problem there is that if I happen to have two classes representing
the same channel, there might be synchronization issues. Maybe I should
have a internal dictionary of created channels? That way, a call to
create a new channel of a given name will return that instance instead
of a new class.